chore: renamed login page
